backend/pyproject.toml
[project]
name = "kathalu-api"
version = "0.1.0"
description = "Backend for the Kathalu Telugu reading app"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "pydantic>=2.7",
  "pydantic-settings>=2.4",
  "python-jose[cryptography]>=3.3",
  "httpx>=0.27",
]

[tool.ruff]
line-length = 100
target-version = "py311"
